cli: Ensure that read timeouts are properly respected
Before, we always updated the timeout before handing the passing the read operation to the connection. This meant that the application could continue to call Read() and could receive new data, although the connection should have timed out. We fix this by only updating the deadline after the read, so repeated calls will not bypass the deadline.